add ucam-webauth-proxy image
Add a simple proxy which can restrict access to a resource to anyone
with a Raven account or to those within a specific lookup group. This is
based on https://github.com/rjw57/ucam-webauth-proxy which is currently
being used to protect our Traefik dashboard.

I've set the version number of the build to 0.9 since we've not hammered
on it much yet and I'd like the freedom to innovate before a 1.0. That
being said, it's deployed already and works so we could bump to 1.0 once
we're happy.

I've also included the ucamwebauth version in the tags in case that ever
becomes an issue for people.